+++ src/components/home/questionnaire-card-content/user-questionnaire-card.tsx
@@ -1,41 +1,26 @@
-import { CardContent, Grid, Skeleton, Typography } from "@mui/material";
-import { useSetAtom, useAtomValue } from "jotai";
-import { useEffect, useState } from "react";
+import { CardContent, Skeleton, Typography } from "@mui/material";
+import { useAtomValue } from "jotai";
 import { userProfileAtom } from "src/atoms/auth";
-import { errorAtom } from "src/atoms/error";
 import { usersAtom } from "src/atoms/user";
-import type { Questionnaire, User } from "src/generated/homeLambdasClient";
-import { useLambdasApi } from "src/hooks/use-api";
+import type { User } from "src/generated/homeLambdasClient";
 import strings from "src/localization/strings";
+import useFetchQuestionnaires from "src/hooks/fetch-questionnaires";
 
+/**
+ * User Questionnaire Card Component
+ * 
+ * @returns User progress in questionnaires
+ */
 const UserQuestionnaireCard = () => {
-  const [loading, setLoading] = useState(false);
-  const [questionnaires, setQuestionnaires] = useState<Questionnaire[]>([]);
-  const { questionnairesApi } = useLambdasApi();
-  const setError = useSetAtom(errorAtom);
+  const { loading, questionnaires } = useFetchQuestionnaires();
   const users = useAtomValue(usersAtom);
   const userProfile = useAtomValue(userProfileAtom);
   const loggedInUser = users.find((user: User) => user.id === userProfile?.id);
-
-  useEffect(() => {
-    const fetchQuestionnaires = async () => {
-      setLoading(true);
-      try {
-        const questionnaires = await questionnairesApi.listQuestionnaires();
-        setQuestionnaires(questionnaires);
-      } catch (error) {
-        setError(`${strings.error.questionnaireLoadFailed}, ${error}`);
-      }
-      setLoading(false);
-    };
-    fetchQuestionnaires();
-  }, []);
 
   /**
    * Count passed questionnaires by Logged in user
    * @returns number of passed questionnaires
    */
-
   const getPassedQuestionnaires = () => {
     if (!loggedInUser) return 0;
     return questionnaires.reduce((count, questionnaire) => {
@@ -52,10 +37,7 @@ const UserQuestionnaireCard = () => {
   const renderUserProgress = () => (
     <>
       <CardContent>
-        <Typography variant="h6" fontWeight={"bold"}>
-          {strings.questionnaireCard.questionnaires}
-        </Typography>
-        <Typography>
+        <Typography variant="body1">
           {strings.formatString(
             strings.questionnaireCard.passedQuestionnaires,
             getPassedQuestionnaires(),
